pyproject.toml
[project]
name = "faa-nasr-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "Paweł Strzelewicz",email = "pawel.strzelewicz83@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "faa_nasr_tools", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pylint (>=4.0.4,<5.0.0)"
]
